export public option types
