+++ packages/react/src/index.ts
@@ -3,5 +3,6 @@ export * from './query';
 export * from './provider';
 export * from './context';
 export * from './tracking';
+export * from './options';
 // re-export the web-sdk so consumers can access that API from the react-sdk
 export * from '@openfeature/web-sdk';
